extract control commands from introspection module
... and put them into a new yoton RepChannel
